feat(#283): create detective
